--- lm_engine/utils/miscellaneous.py
@@ -1,19 +1,0 @@
-# **************************************************
-# Copyright (c) 2026, Mayank Mishra
-# **************************************************
-
-
-def divide_if_divisible(dividend: int, divisor: int, msg: str = "") -> int:
-    """divide if divisible else raise an error
-
-    Args:
-        dividend (int): dividend
-        divisor (int): divisor
-        msg (str): error message
-
-    Returns:
-        int: result
-    """
-
-    assert dividend % divisor == 0, msg
-    return dividend // divisor
